Support env var SLIME_HOME, follow XDG Base Directory SPEC
Support an environment variable named SLIME_HOME to specify the
storage location for data files, currently FASL files.

Since FASL files are generated during compilation, we utilize
XDG_DATA_HOME.

The priority for determining the effective location is as follows:

 1. $SLIME_HOME
 2. $XDG_CACHE_HOME/slime
 3. $HOME/.slime (~/.slime)

This is the subsequent work of slime/slime#610 and slime/slime#787.

Signed-off-by: Burgess Chang <bsc@brsvh.org>